pyproject.toml
[project]
name = "mini-project-mobio"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "duong", email = "vtduong00116@gmail.com"},
]
dependencies = ["mobio-base-sdk==1.0.14", "mobio-admin-sdk>=1.0.15", "python-dateutil>=2.9.0.post0", "m-utilities>=1.0.13", "pydantic>=2.11.4", "m-scheduler-partitioning>=0.1.14", "m-kafka-sdk-v2>=1.0.6", "bson>=0.5.10", "pymongo>=4.12.1", "types-confluent-kafka>=1.3.1", "faker>=37.1.0"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
